added "measures" utility package (with time and space functions) and a new expiry test (which, of course, fails)

+++ DirectMemory-Cache/src/main/java/org/directmemory/CacheEntry.java
@@ -33,7 +33,7 @@ public class CacheEntry implements Comparable<CacheEntry> {
 		return ((expiresOn != null) && Calendar.getInstance().after(expiresOn));
 	}
 	
-	public void expiresIn(int milliseconds) {
+	public void expiresIn(long milliseconds) {
 		if (milliseconds != -1)
 			expiresOn = new Date(new Date().getTime() + milliseconds);
 	}
